Use `NONE` background to match terminal background
I used to set color 0 to the value of the terminal background but this
created problems because applications using this color whould end up
writting in grey on the same grey background.

Instead of setting color 0 to the background color, use `NONE` when
needed which will result in the terminal background color.